Fix "sign up" button css
In my current native resolution, (1366 x 768) the
text would wrap in the bottom sign-up button
despite there being sufficient room for it
visually, so the horizontal padding is replaced with
button width to ensure text has enough horizontal
space to fit in one line. Furthermore, the footer
has been modified as it would cut off at the
bottom of the footer text in Chromium.